added view for airbnbs

+++ App/Controller/FavoritesController.php
@@ -13,34 +13,39 @@ use Laminas\Diactoros\ServerRequest;
 
 class FavoritesController extends Controller
 {
-    public function addEstateToFavorites(ServerRequest $request)
+    public function manageFavoritesByUser(ServerRequest $request)
     {
         $post_data = $request->getParsedBody();
         $user_id = Session::get(Session::USER)->id;
+        $action = $post_data['action'];
+        $estate_id = $post_data['estate_id'];
         $data = [
             'user_id' => $user_id,
-            'estate_id' => $post_data['estate_id'],
+            'estate_id' => $post_data['estate_id']
         ];
-        AppRepoManager::getRm()->getFavoritesRepo()->insertEstateToFavorites($data);
-    }
 
-    public function deleteFavoriteByUser(ServerRequest $request)
-    {
-        $post_data = $request->getParsedBody();
-        $user_id = Session::get(Session::USER)->id;
-        $estate_id = $post_data['estate_id'];
-
-        AppRepoManager::getRm()->getFavoritesRepo()->deleteFavorite($user_id, $estate_id);
+        if ($action == 'add') {
+            AppRepoManager::getRm()->getFavoritesRepo()->insertEstateToFavorites($data);
+        } else if ($action == 'delete') {
+            AppRepoManager::getRm()->getFavoritesRepo()->deleteFavorite($user_id, $estate_id);
+        }
     }
 
     public function seeAllFavoritesByUser(int $id)
     {
         $view_data = [
             'title_tag' => 'Airbnb',
-            'h1_tag' => 'Home page',
+            'h1_tag' => 'Check all your favorites',
             'favorites' => AppRepoManager::getRm()->getFavoritesRepo()->findFavoriteByUserIdWithModels($id),
         ];
         $view = new View('page/wishlist');
         $view->render($view_data);
+    }
+
+    public function deleteFromFavorites(int $id)
+    {
+        $user_id = Session::get(Session::USER)->id;
+        AppRepoManager::getRm()->getFavoritesRepo()->deleteFavoriteByFavoriteId($id);
+        self::redirect('/wishlist/' . $user_id);
     }
 }
